🔨 Move sections out of Stats.vue
