+++ src/sections/TagsSection.vue
@@ -1,0 +1,55 @@
+<template>
+	<div class="tab-area" v-if="display.tags">
+		<SectionTabHeader
+			:tab="tabTags"
+			:tabs="tabsTags"
+			:tab-color-class="() => 'border-bottom-accent3'"
+		/>
+		<div class="tab-content mt-1">
+			<!-- tags count -->
+			<BarChart
+				v-if="tagsChartDataExists"
+				:datasets="tagsChartData.datasets"
+				:labels="tagsChartData.labels"
+				:horizontal="true"
+			/>
+			<div v-if="!tagsChartDataExists" class="tab-empty text-center mt-5">
+				<svg class="icon icon-large icon-gray icon-thin d-block m-0-auto" viewBox="0 0 24 24">
+					<path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+					<path d="M12.983 8.978c3.955 -.182 7.017 -1.446 7.017 -2.978c0 -1.657 -3.582 -3 -8 -3c-1.661 0 -3.204 .19 -4.483 .515m-2.783 1.228c-.471 .382 -.734 .808 -.734 1.257c0 1.22 1.944 2.271 4.734 2.74" />
+					<path d="M4 6v6c0 1.657 3.582 3 8 3c.986 0 1.93 -.067 2.802 -.19m3.187 -.82c1.251 -.53 2.011 -1.228 2.011 -1.99v-6" />
+					<path d="M4 12v6c0 1.657 3.582 3 8 3c3.217 0 5.991 -.712 7.261 -1.74m.739 -3.26v-4" />
+					<line x1="3" y1="3" x2="21" y2="21" />
+				</svg>
+				<div class="text-gray mt-1" v-html="t('stats.charts.tagsCount.empty')"></div>
+			</div>
+		</div>
+	</div>
+</template>
+
+<script setup>
+import { ref, toRef } from 'vue';
+import { useI18n } from 'vue-i18n';
+import { tabsTags } from '@/definitions.js';
+import { useTagsChartData } from '@/composables/useTagsChartData.js';
+import BarChart from '@/charts/BarChart.vue';
+import SectionTabHeader from '@/parts/SectionTabHeader.vue';
+
+const props = defineProps({
+	display: { type: Object, required: true },
+	tags: { type: Array, required: true },
+	options: { type: Object, required: true },
+});
+
+const { t } = useI18n();
+
+// tab navigation local to this section (single-entry tab enum, kept for template consistency)
+const tabTags = ref(tabsTags.tagsCount);
+
+const { tagsChartData, tagsChartDataExists } = useTagsChartData({
+	display: toRef(props, 'display'),
+	tags: toRef(props, 'tags'),
+	options: props.options,
+	t,
+});
+</script>
